requirements.txt
Flask
groq
python-dotenv
numpy
faiss-cpu
sentence-transformers
PyMuPDF
fitz
yfinance
newsapi-python
gunicorn 